% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi-probes.R
\name{.msffi_check}
\alias{.msffi_check}
\title{Re-signal an `msuiter_error_rust` condition surfaced by the Rust core.}
\usage{
.msffi_check(res)
}
\description{
Re-signal an `msuiter_error_rust` condition surfaced by the Rust core.
}
