% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-embed.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Scatter the sample embedding}
\usage{
plot_embedding(embedding)
}
\arguments{
\item{embedding}{The data.frame returned by [ms_embed()].}
}
\value{
A ggplot object.
}
\description{
ggplot scatter of an [ms_embed()] table; colours by cluster when
present.
}
